adjust to never return #f
Avoid a #f return to avoid confusion with defaults from drracket:indentation
(via #lang lines). As #f used to mean "do not move the line", instead return
the current number of spaces. (Note that there are already places where it will
return some specific number of spaces that is equal to the number of spaces
that are already there; the #f was a rare occurrence.)