go back to the old implementation of matching backwards, but keep the cache
as this seems to account for the entire performance difference in modern Racket.
That is, the old version, without the cache, when indenting drracket/private/unit.rkt
(that's already indented) takes about 7% more time to complete than the newer version.
Adding the cache into the older version (as in this commit) recovers that 7%.

(And the older version is simpler.)